+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Exception handler now logs error stacktrace
 - `--endpoint` option for `node init` and `node update` commands
 - `skale info` command with information about current build
+- `skale logs container` command that fetch logs from one of the node containers
 
 ### Changed