+++ CHANGELOG.md
@@ -4,7 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## Unreleased
+
+
+### Added 
+
+- `--manager-url` option for `node init` and `node update` commands 
+- `--ima-url` option for `node init` and `node update` commands 
+- `--dkg-url` option for `node init` and `node update` commands 
+
+### Changed
+
+- `--endpoint` and `--ima-endpoint` options for `node init` and `node update` commands now required
+- `ima` renamed to `mta`
+- Support for the new SKALE node architecture
+
+
+
+## [0.2.0 - 0.3.0] - 2019-09-10
 
 
 ### Added